folders restructured, sass files cleaned up
