+++ public/index.html
@@ -14,8 +14,7 @@
      <link href="https://fonts.googleapis.com/css?family=Abel|Economica|Quicksand" rel="stylesheet">
      <link href="https://fonts.googleapis.com/css?family=Josefin+Sans:600" rel="stylesheet">-->
       <link href='https://fonts.googleapis.com/css?family=Lato:300,400,700,300|Open+Sans:400,600,700,300' rel='stylesheet' type='text/css'>
-      <link href="https://fonts.googleapis.com/css?family=Noto+Serif:400,400italic,700|Open+Sans:300,400,600,700" rel="stylesheet">
-<link href="https://fonts.googleapis.com/css?family=Nanum+Gothic" rel="stylesheet">
+    
     <!-- Latest compiled and minified CSS -->
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
     <!--
