+++ src/App.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import CustomNavbar from './components/CustomNavbar'
+import CustomNavbar from './components/CustomNavbar/CustomNavbar'
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 import Home from './containers/Home';
 import About from './containers/About';
