--- src/components/Carousel.jsx
@@ -1,1 +1,0 @@
--
