+++ src/components/Clients/Clients.jsx
@@ -66,7 +66,7 @@ class Clients extends Component{
         return(
             <div className="item col-xs-6 col-md-3">
             <div className="itemContent">
-            <Image src={require('../images/Clients-logo/'+clientDetails.image+'.png')} alt='' responsive/>
+            <Image src={require('../../images/Clients-logo/'+clientDetails.image+'.png')} alt='' responsive/>
             </div>
             </div>
         );
@@ -80,7 +80,7 @@ class Clients extends Component{
             <div className = "container-fluid"> 
             <div className = "container">
             <div className = "title col-sm-12 col-md-3">
-            <div className ="row"><h1>Clients</h1> </div>
+            <div className ="row"><h1><strong>Clients</strong></h1> </div>
             <div className ="content">It is primarily used for source code management in software dev.but it can be used to keep track changes in set of files</div>
             </div>
             <div className = "grid col-sm-12 col-md-9">
