--- src/components/EmbeddedServices.css
@@ -1,42 +1,0 @@
-h1, h2, h3, h4, h5, h6 {
-  font-family: 'Lato', sans-serif;
-  font-weight: 300; }
-
-.show-grid-fluid-embedded-services {
-  background-color: #f7f7f7; }
-
-.show-grid-embedded-services {
-  background-color: #f7f7f7; }
-
-.show-title-col {
-  background-color: #f7f7f7;
-  color: black;
-  padding-top: 10px; }
-
-.show-title-col-row-title {
-  margin: 10px;
-  margin-top: 15px;
-  border-left: 3px solid;
-  Padding-left: 15px;
-  padding-top: 0px; }
-
-.show-title-col-row-content {
-  margin: 10px;
-  text-align: justify;
-  font-family: "Open Sans", Arial, sans-serif;
-  font-size: 14px;
-  font-weight: 100; }
-
-.show-col-grid {
-  background-color: #f7f7f7; }
-
-.show-col-grid-col {
-  background-color: #f7f7f7; }
-
-.show-col-grid-col-row-item {
-  background-color: #112a39 !important;
-  height: 300px;
-  margin-top: 20px;
-  margin-left: 0px;
-  margin-right: 0px;
-  margin-bottom: 20px; }
