+++ src/components/Footer/Footer.css
@@ -1,4 +1,5 @@
 /*colors*/
+/*primary colors */
 h1, h2, h3, h4, h5, h6 {
   font-family: 'Lato', sans-serif;
   font-weight: 300; }
@@ -56,7 +57,7 @@ h1, h2, h3, h4, h5, h6 {
   padding-right: 30px;
   padding-bottom: 90px; }
   .footer-style .footer-column h1, .footer-style .footer-column h2, .footer-style .footer-column h3, .footer-style .footer-column h4, .footer-style .footer-column h5, .footer-style .footer-column h6 {
-    font-family: 'Lato', sans-serif;
+    font-family: "Lato", sans-serif;
     font-weight: 300;
     margin-bottom: 10px;
     color: #f7f7f7; }
