+++ src/components/Modal/Modal.css
@@ -1,12 +1,13 @@
 /*colors*/
+/*primary colors */
 h1, h2, h3, h4, h5, h6 {
   font-family: 'Lato', sans-serif;
-  font-weight: 400; }
+  font-weight: 300; }
 
 .modal-content {
   border-radius: 0px; }
   .modal-content .btn-modal {
-    background-color: #112a39;
+    background-color: #1c2331;
     color: #fff;
     border-radius: 0px;
     -webkit-transition: .5s ease;
