+++ src/components/Services/OurEmbeddedServices.jsx
@@ -2,15 +2,15 @@ import React,{ Component } from 'react';
 import { Grid, Row, Col, Glyphicon} from 'react-bootstrap';
 import './OurServices.css';
 /* import icons*/
-import MdPhonelink from 'react-icons/lib/md/phonelink';
+import FaCogs from 'react-icons/lib/fa/cogs';
 import MdDeveloperMode from 'react-icons/lib/md/developer-mode';
-import FaCartArrowDown from 'react-icons/lib/fa/cart-arrow-down';
+import FaCloud from 'react-icons/lib/fa/cloud';
 import FaSearchPlus from 'react-icons/lib/fa/search-plus';
 import FaServer from 'react-icons/lib/fa/server';
 import FaUserPlus from 'react-icons/lib/fa/user-plus';
 
 
-const OurServices = ()=>{
+const OurEmbeddedServices = ()=>{
     
     return(
     <div>
@@ -21,7 +21,7 @@ const OurServices = ()=>{
     <Row className="rowsTile">
        <Col xs ={12} md={12} className="columnText">
         <Row className ="titleRow">
-            <h1>Software Services</h1>
+            <h1><strong>Embedded Services </strong></h1>
         </Row>
         <Row className ="titleRowContent">
            We spealize in designing fully responsive web applications using modern day framworks.We are a leading service provider in the digital domain, an organization that aspires to drive digital transformation. With our expertise ranging from custom product development, enterprise mobility, UI/UX design, digital marketing, and e-commerce, to technologies such as AI, AR, VR, RPA, and more, we emerge as the one-stop destination for the entire IT needs of businesses. We provide futuristic digital transformation services that enable businesses to keep pace with the emerging trends.
@@ -34,9 +34,9 @@ const OurServices = ()=>{
     </Col>
     <Col xs={12} md={4} className="columns">
     <Row className="rows">
-        <Col xs ={4} md={4}  ><MdPhonelink className = "icon1"/></Col>
+        <Col xs ={4} md={4}  ><FaCogs className = "icon1"/></Col>
         <Col xs ={8} md={8} className="coltext">
-        <h3>Web Development</h3>
+        <h3>Product Development</h3>
             We spealize in designing fully responsive web applications using modern day frameworks.
         </Col>
         </Row>
@@ -48,9 +48,9 @@ const OurServices = ()=>{
         </Col>
         </Row>
     <Row className="rows">
-        <Col xs ={4} md={4}  >< FaCartArrowDown  className = "icon1"/></Col>
+        <Col xs ={4} md={4}  >< FaCloud  className = "icon1"/></Col>
         <Col xs ={8} md={8} className="coltext">
-        <h3>E-Commerce</h3>
+        <h3>IoT Products</h3>
           How many of your website visitors are turning into customers? We will help make your ecommerce website your most effective selling tool.
         </Col>
         </Row>
@@ -84,4 +84,4 @@ const OurServices = ()=>{
     </div>
     );
 }
-export default OurServices;
+export default OurEmbeddedServices;
