--- src/components/Slider.css
@@ -1,50 +1,0 @@
-.tech{
-    background-color: aqua;
-}
-.techHeading{
-    
-    color:#fff;
-    border-left:3px solid white;
-    margin: 10px;
-}
-.list-pic{
-    
-    height:170px;
-    background-color: transparent;
-    list-style: none;
-    padding:10px;
-}
-.list-pic-innerBlock{
-    height:150px;
-    background-color:#112a39!important;;
-    text-align: center;
-    padding-top:25px;
-    padding-bottom: 25px;
-}
-.overlay {
-  position: absolute;
-  top: 10px;
-  bottom: 10px;
-  left: 10px;
-  right: 10px;
-  height: 150px;
-  width:  90%;
-  opacity: 0;
-  transition: .5s ease;
-  background-color: #f5b929;
-}
-.text {
-  color: black;
-  font-size: 20px;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  -ms-transform: translate(-50%, -50%);
-  text-align: center;
-}
-.list-pic-innerBlock:hover .overlay{
-  
-    opacity:0.75;
-}
-
