--- src/components/Slider.jsx
@@ -1,94 +1,0 @@
-import React, { Compoenent } from 'react'
-import { Carousel, Grid, Row, Col, Glyphicon,Images} from 'react-bootstrap';
-import "./Slider.css";
-const tech =()=>{
-
-return(
-<div>
-  <div className = "container">
-  <div className ="row">
-  <div className = "col-xs-12 col-md-12">
-        <div className = 'techHeading col-md-12'>
-            <h1>Technological Expertize</h1>
-        </div>
-  </div>
-  </div>
-  <div className ="row">
-      <div className ="list-pic col-xs-6 col-md-4 col-lg-2">
-        <div className = "list-pic-innerBlock">
-        <img src={require('../images/react.png')} alt=''/>
-        <div class="overlay">
-        <div class="text">React.Js</div>
-        </div>
-        </div>
-      </div>
-       <div className ="list-pic col-xs-6 col-md-4 col-lg-2">
-        <div className = "list-pic-innerBlock">
-        <img src={require('../images/react.png')} alt=''/>
-        </div>
-      </div> 
-       <div className ="list-pic col-xs-6 col-md-4 col-lg-2">
-        <div className = "list-pic-innerBlock">
-        <img src={require('../images/react.png')} alt=''/>
-        </div>
-      </div> 
-       <div className ="list-pic col-xs-6 col-md-4 col-lg-2">
-        <div className = "list-pic-innerBlock">
-        <img src={require('../images/react.png')} alt=''/>
-        </div>
-      </div> 
-       <div className ="list-pic col-xs-6 col-md-4 col-lg-2">
-        <div className = "list-pic-innerBlock">
-        <img src={require('../images/react.png')} alt=''/>
-        </div>
-      </div> 
-       <div className ="list-pic col-xs-6 col-md-4 col-lg-2">
-        <div className = "list-pic-innerBlock">
-        <img src={require('../images/react.png')} alt=''/>
-        </div>
-      </div> 
-      
-  </div>
-   <div className ="row">
-      <div className ="list-pic col-xs-6 col-md-4 col-lg-2">
-        <div className = "list-pic-innerBlock">
-        <img src={require('../images/react.png')} alt=''/>
-        </div>
-      </div>
-       <div className ="list-pic col-xs-6 col-md-4 col-lg-2">
-        <div className = "list-pic-innerBlock">
-        <img src={require('../images/react.png')} alt=''/>
-        
-        </div>
-      </div> 
-       <div className ="list-pic col-xs-6 col-md-4 col-lg-2">
-        <div className = "list-pic-innerBlock">
-        <img src={require('../images/react.png')} alt=''/>
-        </div>
-      </div> 
-       <div className ="list-pic col-xs-6 col-md-4 col-lg-2">
-        <div className = "list-pic-innerBlock">
-        <img src={require('../images/react.png')} alt=''/>
-        </div>
-      </div> 
-       <div className ="list-pic col-xs-6 col-md-4 col-lg-2">
-        <div className = "list-pic-innerBlock">
-        <img src={require('../images/react.png')} alt=''/>
-        </div>
-      </div> 
-       <div className ="list-pic col-xs-6 col-md-4 col-lg-2">
-        <div className = "list-pic-innerBlock">
-        <img src={require('../images/react.png')} alt=''/>
-        </div>
-      </div> 
-  </div>    
-  
-  </div>
-  
-
-</div>
-
-);
-}
-export default tech;
-
