+++ src/components/Subscribe/Subscribe.css
@@ -1,25 +1,30 @@
 h1, h2, h3, h4, h5, h6 {
   font-family: 'Lato', sans-serif;
-  font-weight: 300; }
+  font-weight: 400;
+  margin-bottom: 20px; }
 
+/*colors*/
+/*primary colors */
 .baseGrid {
-  height: 200px;
   background-color: transparent;
   color: #fff; }
-
-.subscribeText {
-  padding-top: 50px; }
+  .baseGrid .show-grid .subscribeText {
+    padding-top: 50px;
+    padding-bottom: 50px; }
+  .baseGrid .show-grid .subscribeButtonHolder {
+    padding-top: 80px;
+    padding-bottom: 50px; }
 
 .btn-primary2 {
   color: #fff;
+  margin: 0 auto;
   float: right;
   border-radius: 0px;
-  background-color: #112a39;
-  border-color: #4c86a3;
-  margin-top: 80px;
+  background-color: #1c2331;
+  border-color: #f5b929;
   width: 200px;
   height: 50px;
-  font-family: "Open Sans", Arial, sans-serif;
+  font-family: "Open Sans", sans-serif;
   font-size: 18px;
   font-weight: 300;
   -webkit-transition: background-color 1s ease-in;
@@ -27,31 +32,37 @@ h1, h2, h3, h4, h5, h6 {
   -o-transition: background-color 1s ease-in;
   transition: background-color 1s ease-in; }
   .btn-primary2:hover {
-    background-color: #0d202c;
-    color: #000;
-    background-color: #ededed;
+    background-color: #f7f7f7;
+    color: #696969;
     -webkit-transition: background-color 1s ease-out;
     -moz-transition: background-color 1s ease-out;
     -o-transition: background-color 1s ease-out;
     transition: background-color 1s ease-out; }
 
 @media (max-width: 767px) {
-  .baseGrid {
-    height: 350px;
-    background-color: transparent;
-    color: #fff; }
-  .subscribeText {
-    padding-top: 50px; }
+  .baseGrid .show-grid .subscribeText {
+    padding-top: 50px;
+    padding-bottom: 50px; }
+  .baseGrid .show-grid .subscribeButtonHolder {
+    padding-top: 0px;
+    padding-bottom: 20px;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    -webkit-box-pack: center;
+    -ms-flex-pack: center;
+    justify-content: center; }
   .btn-primary2 {
     color: #fff;
-    float: right;
     border-radius: 0px;
-    background-color: #112a39;
-    border-color: #4c86a3;
-    margin-top: 10px;
+    background-color: #1c2331;
+    border-color: #f5b929;
     width: 200px;
     height: 50px;
-    font-family: "Open Sans", Arial, sans-serif;
+    font-family: "Open Sans", sans-serif;
     font-size: 18px;
     font-weight: 300;
     -webkit-transition: background-color 1s ease-in;
@@ -59,9 +70,8 @@ h1, h2, h3, h4, h5, h6 {
     -o-transition: background-color 1s ease-in;
     transition: background-color 1s ease-in; }
     .btn-primary2:hover {
-      background-color: #0d202c;
-      color: #000;
-      background-color: #ededed;
+      background-color: #f7f7f7;
+      color: #696969;
       -webkit-transition: background-color 1s ease-out;
       -moz-transition: background-color 1s ease-out;
       -o-transition: background-color 1s ease-out;
