+++ src/components/Subscribe/Subscribe.jsx
@@ -13,10 +13,10 @@ return(
     <Row className="show-grid">
    
     <Col xs={12} md={8} className ="subscribeText">
-      <h1> Subscribe Us - For Updates ,News & Offers</h1>
+      <h1><strong>Subscribe Us - For Updates ,News & Offers</strong> </h1>
       <p>Join our 1000+ subscribers and get access to the latest tools, freebies, product announcements and much more!</p>
     </Col>
-    <Col xs={12} md={4}>
+    <Col xs={12} md={4} className ="subscribeButtonHolder">
       <Button  bsStyle="primary2">SUBSCRIBE</Button>
     </Col>
   </Row>
