+++ src/components/Technologies/Tech.css
@@ -2,7 +2,8 @@
 /*primary colors */
 h1, h2, h3, h4, h5, h6 {
   font-family: 'Lato', sans-serif;
-  font-weight: 300; }
+  font-weight: 400;
+  margin-bottom: 20px; }
 
 .techContainer {
   margin-top: 50px;
@@ -23,14 +24,15 @@ h1, h2, h3, h4, h5, h6 {
 
 .list-pic-innerBlock {
   height: 150px;
-  background-color: #f7f7f7;
+  background-color: white;
   -webkit-box-shadow: 0 1px 21px rgba(0, 0, 0, 0.19);
   box-shadow: 0 1px 21px rgba(0, 0, 0, 0.19);
   text-align: center;
   padding-top: 25px;
   padding-bottom: 25px; }
   .list-pic-innerBlock:hover .overlay {
-    opacity: 0.85; }
+    opacity: 0.95;
+    border: 1px solid #f5b929; }
   .list-pic-innerBlock .overlay {
     position: absolute;
     top: 10px;
@@ -50,8 +52,8 @@ h1, h2, h3, h4, h5, h6 {
     color: #fff;
     font-size: 12px;
     position: absolute;
-    text-align: left;
-    font-family: "Open Sans", Arial, sans-serif;
+    text-align: center;
+    font-family: "Open Sans", sans-serif;
     padding: 15px; }
     .list-pic-innerBlock .text h2 {
       text-align: center; }
