--- src/components/WhatWeDo.css
@@ -1,43 +1,0 @@
-.item{
-    
-    /*background-color:#f4bc1f !important ;*/
-    background-color:#c4c4c4 !important ;
-    height: 325px;
-    text-align: justify;
-    text-justify: inter-word;
-    margin-top: 10px;
-    border-radius:10px;
-    border-color: aliceblue;
-    padding:25px;
-    color:black;
-}
-
-.show-grid{
-    color:black;
-    background-color: #c4c4c4!important;
-    text-align: center;
-    font-family: 'Abel', sans-serif;
-    height: 300px;
-    text-align: center;
-}
-.iconHolder
-{
-    text-align: center;
-}
-.icon{
-    font-size: 100px; 
-}
-
-.title
-{
-    text-align: center;  
-    
-}
-.outterRing
-{
-    background-color: #c4c4c4!important;
-}
-.line{
-    height:1px;
-    background-color:black;
-}
