+++ src/components/tech.css
@@ -1,7 +1,7 @@
 /*colors*/
 h1, h2, h3, h4, h5, h6 {
   font-family: 'Lato', sans-serif;
-  font-weight: 400; }
+  font-weight: 300; }
 
 .techContainer {
   margin-top: 50px;
