+++ src/containers/Home.jsx
@@ -1,13 +1,12 @@
 import React, { Component } from 'react';
-import JumbotronFront from '../components/Jumbotron';
-import Intro from '../components/intro';
-import OurServices from '../components/OurServices';
-import Subscribe from '../components/Subscribe';
-import EmbeddedServices from '../components/EmbeddedServices';
-import OurEmbeddedServices from '../components/OurEmbeddedServices';
-import Tech from '../components/tech';
-import Clients from '../components/Clients';
-import Footer from '../components/Footer';
+import JumbotronFront from '../components/Jumbotron/Jumbotron';
+import Intro from '../components/Intro/Intro';
+import OurServices from '../components/Services/OurServices';
+import Subscribe from '../components/Subscribe/Subscribe';
+import OurEmbeddedServices from '../components/Services/OurEmbeddedServices';
+import Tech from '../components/Technologies/Tech';
+import Clients from '../components/Clients/Clients';
+import Footer from '../components/Footer/Footer';
 import './Home.css';
 
 class Home extends Component
