+++ src/css/components/Jumbotron.css
@@ -1,8 +1,17 @@
+/*colors*/
+/*primary colors */
 h1, h2, h3, h4, h5, h6 {
   font-family: 'Lato', sans-serif;
   font-weight: 300; }
 
-/*colors*/
+/*$themeColor : #1c2331;
+$themeDark  : darken($themeColor,2%);
+$pageColor  : #f7f7f7;
+$decorativeColor: #f5b929;
+$linkColor:#bfbfbf;
+$textColorLight:#fff;
+$textColorDark:#fff;
+$smallLine:lighten(#1c2331, 40% );*/
 .jumbotron {
   background: transparent;
   font-family: "Lato", sans-serif;
@@ -16,8 +25,8 @@ h1, h2, h3, h4, h5, h6 {
   .jumbotron .btn-primary {
     color: #fff;
     border-radius: 0px;
-    background-color: #112a39;
-    border-color: #4c86a3;
+    background-color: #1c2331;
+    border-color: #f5b929;
     margin-left: 10px;
     width: 150px;
     height: 50px;
@@ -30,7 +39,8 @@ h1, h2, h3, h4, h5, h6 {
     transition: background-color 1s ease-in; }
     .jumbotron .btn-primary:hover {
       color: #000;
-      background-color: #ededed;
+      border-color: #f5b929;
+      background-color: #f7f7f7;
       -webkit-transition: background-color 1s ease-out;
       -moz-transition: background-color 1s ease-out;
       -o-transition: background-color 1s ease-out;
@@ -44,8 +54,8 @@ h1, h2, h3, h4, h5, h6 {
     margin-bottom: 0px; }
     .jumbotron .btn-primary {
       color: #fff;
-      background-color: #112a39;
-      border-color: #4c86a3;
+      background-color: #1c2331;
+      border-color: #f5b929;
       margin-left: 0px;
       width: 150px;
       height: 50px;
@@ -57,7 +67,7 @@ h1, h2, h3, h4, h5, h6 {
       transition: background-color 1s ease-in; }
       .jumbotron .btn-primary:hover {
         color: #000;
-        background-color: #ededed;
+        background-color: #f7f7f7;
         -webkit-transition: background-color 1s ease-out;
         -moz-transition: background-color 1s ease-out;
         -o-transition: background-color 1s ease-out;
