+++ src/sass/Colors.scss
@@ -1,17 +1,10 @@
 /*colors*/
-$themeColor : #112a39!important;
-$hoverColor :#fff;
-$overlayColor :#f5b929;
-$overlayText:#000;
-$textColorLight:#fff;
-$textColorDark:#000;
-$buttonBorder :#4c86a3;
-$buttonHover:#ededed;
-$buttonColor:#112a39;
-$pageColor:#f7f7f7;
-$smallLine:#415d6e;
-$footerbg:#1c2331;
-$footerbgHigh:#161c27;
+/*primary colors */
+$themeColor : #1c2331;
+$themeDark  : darken($themeColor,2%);
+$pageColor  : #f7f7f7;
+$decorativeColor: #f5b929;
 $linkColor:#bfbfbf;
-$linkHover:#fff;
-$textLine:#f5b929;
+$textColorLight:#fff;
+$textColorDark:#696969;
+$smallLine:lighten(#1c2331, 40% );
