--- src/sass/components/EmbeddedServices.scss
@@ -1,50 +1,0 @@
-@import "../Variables";
-@import "../mixins";
-
-.show-grid-fluid-embedded-services{
-      background-color: $pageColor
-}
-.show-grid-embedded-services{
-      background-color: $pageColor;
-}
-.show-title-col{
-      background-color: $pageColor;
-      color:black;
-      padding-top:10px;
-}
-.show-title-col-row-title
-{
-    margin:10px;
-    margin-top:15px;
-    border-left:3px solid ;
-    Padding-left: 15px;
-    padding-top:0px;
-    
-   
-}
-.show-title-col-row-content
-{
-    margin:10px;
-    text-align: justify;
-    font-family: $font;
-    font-size:14px;
-    font-weight:100;
-    
-}
-.show-col-grid{
-      background-color: $pageColor;
-      //height: 800px;
-}
-.show-col-grid-col
-{
-    background-color: $pageColor;
-    //height: 875px;
-}
-.show-col-grid-col-row-item{
-    background-color: $themeColor;
-    height:300px;
-    margin-top:20px;
-    margin-left:0px;
-    margin-right:0px;
-    margin-bottom:20px;
-}
