+++ src/sass/components/Modal.scss
@@ -1,5 +1,5 @@
-@import "../Colors";
-@import "../Fonts";
+@import "../colors";
+@import "../fonts";
 @import "../mixins";
 .modal-content{
     border-radius:0px;
