+++ src/sass/fonts.scss
@@ -1,6 +1,7 @@
-$font:'Open Sans', Arial, sans-serif;
+$font:'Open Sans',sans-serif;
 $fontHighLight: 'Lato', sans-serif;
 h1, h2, h3, h4, h5, h6 {
   font-family: 'Lato', sans-serif;
-  font-weight: 300;
+  font-weight: 400;
+  margin-bottom: 20px;
 }
